pyproject.toml
[tool.poetry]
name = "discord"
version = "0.1.0"
description = ""
authors = ["sarushili0430 <38129719+sarushili0430@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^1.7.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
